Move Account to shared/models
