+++ client/src/app/core/authentication.service.ts
@@ -3,7 +3,7 @@ import {HttpClient, HttpErrorResponse} from "@angular/common/http";
 import {Injectable} from "@angular/core";
 import {environment} from "../../environments/environment";
 import {CoreSecurity} from "./security.module";
-import {Account} from "./security.data";
+import {Account} from "../shared/models/account";
 
 @Injectable({
   providedIn: CoreSecurity,
